fixed helpers; added image.h

+++ lib/lzss/lzss.h
@@ -35,11 +35,11 @@ static uint32_t head[1<<24-1];   /* head of chain for each hash */
 static uint32_t prev[1<<24-1]; /* previous positions in the same hash */
 
 extern uint32_t hash3(const uint8_t *buf);
-extern void insert_hash(const uint8_t* window, int pos);
+extern void insert_hash(const uint8_t* window, long int pos);
 
-extern int find_match(const uint8_t* window, int pos, int maxlen, int *best_dist);
+extern int find_match(const uint8_t* window, long int pos, int maxlen, int *best_dist);
 /*return ptr count*/
-extern int generate_lzss_pointers(uint8_t* input,int input_size,slzss_pointer* lzss_ptrs,int ptr_n,int* data_cnt,
+extern int generate_lzss_pointers(uint8_t* input,long int input_size,slzss_pointer* lzss_ptrs,int ptr_n,long int* data_cnt,
                             f_on_emit_lzss_ptr fn_olptr,
                             f_count_distinct_literal fn_literalcount
                             );
